splitting up and renaming utilities in util
